Refactor Link component and related files
