+++ app/modules/OrderCard.tsx
@@ -1,8 +1,8 @@
 import { Image, flattenConnection } from "@shopify/hydrogen";
-
 import type { OrderCardFragment } from "customer-accountapi.generated";
 import { statusMessage } from "~/lib/utils";
-import { Heading, Link, Text } from "~/modules";
+import { Heading, Text } from "~/modules";
+import { Link } from "~/components/Link";
 
 export function OrderCard({ order }: { order: OrderCardFragment }) {
   if (!order?.id) return null;
