+++ app/routes/($locale).account.orders.$id.tsx
@@ -3,11 +3,11 @@ import { Image, Money, flattenConnection } from "@shopify/hydrogen";
 import { type LoaderFunctionArgs, json, redirect } from "@shopify/remix-oxygen";
 import clsx from "clsx";
 import invariant from "tiny-invariant";
-
 import type { OrderFragment } from "customer-accountapi.generated";
 import { CUSTOMER_ORDER_QUERY } from "~/graphql/customer-account/CustomerOrderQuery";
 import { statusMessage } from "~/lib/utils";
-import { Heading, Link, PageHeader, Text } from "~/modules";
+import { Heading, PageHeader, Text } from "~/modules";
+import { Link } from "~/components/Link";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return [{ title: `Order ${data?.order?.name}` }];
